Adds an editor (#25)
This adds a basic editor and allows imports of rides from the editor.
There's lots to improve here, so it is just a first draft and will need to be thoroughly tested in production. Currently, it works fine in staging, but because of the login mechanisms requiring certain URLs, the preview link freezes up a lot. In the future, we'll have to figure out how to have preview links work with the OAuth2 flow.